Split Gtlink out from chain module to split out ST dependency
