+++ fermipy/diffuse/job_library.py
@@ -7,7 +7,8 @@ from __future__ import absolute_import, division, print_function
 import os
 import sys
 
-from fermipy.jobs.chain import Gtlink, Link
+from fermipy.jobs.chain import Link
+from fermipy.jobs.gtlink import Gtlink
 from fermipy.jobs.scatter_gather import ConfigMaker
 from fermipy.jobs.lsf_impl import build_sg_from_link
 from fermipy.diffuse.name_policy import NameFactory
